# backend-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NewAPI Tools 增强管理中间件 API 文档",
        "title": "NewAPI Tools API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/linuxdo/lookup/{linux_do_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LinuxDo"
                ],
                "summary": "查询 LinuxDo 用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LinuxDo 用户 ID",
                        "name": "linux_do_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/status-page/heartbeat/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "状态页"
                ],
                "summary": "获取状态页心跳数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态页标识",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "时间窗口",
                        "name": "window",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/status-page/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "状态页"
                ],
                "summary": "获取状态页配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态页标识",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "时间窗口",
                        "name": "window",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/status-page/{slug}/badge": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "状态页"
                ],
                "summary": "获取状态页徽章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态页标识",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "时间窗口",
                        "name": "window",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "徽章标签",
                        "name": "label",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/status-page/{slug}/summary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "状态页"
                ],
                "summary": "获取状态页摘要",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态页标识",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "时间窗口",
                        "name": "window",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/tokens": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "令牌管理"
                ],
                "summary": "获取令牌列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态过滤 (active/disabled/expired)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称过滤",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tokens/statistics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "令牌管理"
                ],
                "summary": "获取令牌统计信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}